refactor: Differentiate dto response request folder
