+++ src/main/java/com/psj/itembrowser/order/service/OrderService.java
@@ -1,6 +1,6 @@
 package com.psj.itembrowser.order.service;
 
-import com.psj.itembrowser.order.domain.dto.OrderResponseDTO;
+import com.psj.itembrowser.order.domain.dto.response.OrderResponseDTO;
 
 public interface OrderService {
